Complete preliminary implementation of ABS test

+++ src/models/patient.tsx
@@ -9,7 +9,7 @@ import {
     deleteDoc,
 } from "firebase/firestore";
 
-import { v4 as generateId, validate as validateId } from "uuid";
+import { validateId } from "@/models";
 
 export type PatientData = {
     createdOn: Timestamp;
@@ -77,60 +77,11 @@ export async function deletePatient(
     database: FirebaseFirestore,
     patientId: string
 ) {
+    if (!validateId(patientId)) {
+        throw "Invalid patient identifier";
+    }
+
     const path = getPatientPath(patientId);
     const document = doc(database, path);
     await deleteDoc(document);
 }
-
-export type AbsData = {};
-
-function getAbsPath(patientId: string, testId: string) {
-    return `patients/${patientId}/abs/${testId}`;
-}
-
-// async function getAbs(database: Database, patientId: string, testId: string) {
-//     const path = getAbsPath(patientId, testId);
-//     const document = doc(database, path);
-//     await setDoc<AbsData>(document, {});
-// }
-
-// OLD
-
-export type Patient = {
-    createdOn: string;
-    lastAccessedOn: string;
-};
-
-export enum AbsResponse {
-    Absent = 1,
-    Slight = 2,
-    Moderate = 3,
-    Extreme = 4,
-}
-
-export type AbsTest = {
-    environment: string;
-
-    period: {
-        from: Timestamp;
-        to: Timestamp;
-    };
-
-    responses: AbsResponse[];
-
-    score: {
-        disinhibition: number;
-        aggression: number;
-        lability: number;
-    };
-};
-
-export enum WptasResponse {
-    Correct = 1,
-    Incorrect = 2,
-    CorrectPrompted = 3,
-}
-
-export type WptasTest = {
-    responses: WptasResponse[];
-};
